updates: migrations to ensure drop/create tables are being migrated safely

+++ db/migrate/20140203145644_create_spree_bit_credits_checkouts.rb
@@ -1,16 +1,16 @@
 class CreateSpreeBitCreditsCheckouts < ActiveRecord::Migration
   def up
-    return if ActiveRecord::Base.connection.table_exists? 'spree_bit_credits_checkouts'
-
-    create_table :bit_credits_checkouts do |t|
-      t.float :old_balance
-      t.float :new_balance
-      t.string :source
-      t.integer :tx_id
+    unless ActiveRecord::Base.connection.table_exists? 'spree_bit_credits_checkouts'
+      create_table :bit_credits_checkouts do |t|
+        t.float :old_balance
+        t.float :new_balance
+        t.string :source
+        t.integer :tx_id
+      end
     end
   end
 
   def down
-    drop_table :bit_credits_checkouts
+    drop_table :bit_credits_checkouts if ActiveRecord::Base.connection.table_exists? 'spree_bit_credits_checkouts'
   end
 end
